rename composition file name to bodycomposition
